requirements.txt
kafka-python
requests
pyspark
streamlit
pandas
scikit-learn
joblib
numpy
